# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Oddiya Travel Planning API
  description: |
    AI-powered travel planning and video generation service API.
    
    ## Features
    - OAuth2 authentication (Google, Apple)
    - AI-powered travel plan generation using AWS Bedrock
    - Photo to video shorts generation
    - Place discovery and recommendations
    - Real-time notifications
    
    ## Authentication
    Most endpoints require Bearer token authentication.
    Include the token in the Authorization header: `Bearer {accessToken}`
  version: 1.0.0
  contact:
    name: Oddiya API Support
    email: api-support@oddiya.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.oddiya.com/api/v1
    description: Production server
  - url: https://api-test.oddiya.com/api/v1
    description: Test server
  - url: http://localhost:8080/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with OAuth provider token
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate user session and tokens
      operationId: logoutUser
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /travel-plans/generate:
    post:
      tags:
        - Travel Plans
      summary: Generate AI travel plan
      description: Generate travel plan using AI based on user preferences
      operationId: generateTravelPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneratePlanRequest'
      responses:
        '201':
          description: Travel plan generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravelPlanDetailed'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /places/search:
    get:
      tags:
        - Places
      summary: Search places
      description: Search for places by query and filters
      operationId: searchPlaces
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          description: Search query
        - name: city
          in: query
          schema:
            type: string
          description: City filter
        - name: category
          in: query
          schema:
            type: string
            enum: [attraction, restaurant, accommodation, shopping]
          description: Place category
        - name: latitude
          in: query
          schema:
            type: number
            format: double
          description: Center latitude for proximity search
        - name: longitude
          in: query
          schema:
            type: number
            format: double
          description: Center longitude for proximity search
        - name: radius
          in: query
          schema:
            type: integer
            default: 5000
          description: Search radius in meters
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Places found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /places/{placeId}:
    get:
      tags:
        - Places
      summary: Get place details
      description: Get detailed information about a specific place
      operationId: getPlaceDetails
      parameters:
        - name: placeId
          in: path
          required: true
          schema:
            type: string
          description: Place ID
      responses:
        '200':
          description: Place details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Results per page
...
  schemas:
...
    LoginRequest:
      type: object
      required:
        - provider
        - idToken
      properties:
        provider:
          type: string
          enum: [google, apple]
        idToken:
          type: string
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
...
    GeneratePlanRequest:
      type: object
      required:
        - destination
        - startDate
        - endDate
        - budget
      properties:
        destination:
          type: string
          description: Travel destination
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: number
          minimum: 0
          description: Budget in KRW
        companions:
          $ref: '#/components/schemas/Companions'
        preferences:
          $ref: '#/components/schemas/TravelPreferences'
        accommodationPreference:
          type: string
          enum: [hotel, guesthouse, airbnb]
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        error:
          $ref: '#/components/schemas/ErrorDetail'
          nullable: true
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
        timestamp:
          type: string
          format: date-time
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        version:
          type: string
    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: '#/components/schemas/User'
                tokens:
                  $ref: '#/components/schemas/Tokens'
    TokenResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Tokens'
...
    PlaceSearchResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/PlaceSummary'
                pagination:
                  $ref: '#/components/schemas/Pagination'
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        nickname:
          type: string
        profileImageUrl:
          type: string
          format: uri
        travelStyle:
          type: string
          enum: [adventurous, relaxed, cultural, luxury]
        createdAt:
          type: string
          format: date-time
...
    Tokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Expiration time in seconds
...
    TravelPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        destination:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: number
        status:
          type: string
          enum: [draft, confirmed, completed, cancelled]
        createdAt:
          type: string
          format: date-time
...
    TravelPlanDetailed:
      allOf:
        - $ref: '#/components/schemas/TravelPlan'
        - type: object
          properties:
            companions:
              $ref: '#/components/schemas/Companions'
            dailyPlans:
              type: array
              items:
                $ref: '#/components/schemas/DailyPlan'
            participants:
              type: array
              items:
                $ref: '#/components/schemas/Participant'
            statistics:
              $ref: '#/components/schemas/PlanStatistics'
            metadata:
              $ref: '#/components/schemas/PlanMetadata'
            updatedAt:
              type: string
              format: date-time
    Companions:
      type: object
      properties:
        adults:
          type: integer
          minimum: 1
          default: 1
        children:
          type: integer
          minimum: 0
          default: 0
    TravelPreferences:
      type: object
      properties:
        themes:
          type: array
          items:
            type: string
            enum: [culture, food, shopping, nature, adventure, relaxation]
        pace:
          type: string
          enum: [slow, moderate, fast]
        mustVisit:
          type: array
          items:
            type: string
    DailyPlan:
      type: object
      properties:
        day:
          type: integer
        date:
          type: string
          format: date
        title:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryItem'
        estimatedCost:
          type: number
        totalDistance:
          type: number
          description: Total distance in km
    ItineraryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sequence:
          type: integer
        time:
          type: string
        placeName:
          type: string
        placeId:
          type: string
        category:
          type: string
        duration:
          type: integer
          description: Duration in minutes
        cost:
          type: number
        description:
          type: string
        location:
          $ref: '#/components/schemas/Location'
        tips:
          type: string
    Location:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        address:
          type: string
...
    Participant:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        nickname:
          type: string
        role:
          type: string
          enum: [owner, participant]
        joinedAt:
          type: string
          format: date-time
    PlanStatistics:
      type: object
      properties:
        totalPlaces:
          type: integer
        totalDistance:
          type: number
        estimatedCost:
          type: number
    PlanMetadata:
      type: object
      properties:
        generatedBy:
          type: string
          enum: [ai, manual]
        modelVersion:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
    PlaceSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        placeId:
          type: string
        name:
          type: string
        category:
          type: string
        rating:
          type: number
          minimum: 0
          maximum: 5
        reviewCount:
          type: integer
        priceLevel:
          type: integer
          minimum: 1
          maximum: 4
        location:
          $ref: '#/components/schemas/Location'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        openingHours:
          $ref: '#/components/schemas/OpeningHours'
        tags:
          type: array
          items:
            type: string
    PlaceDetails:
      allOf:
        - $ref: '#/components/schemas/PlaceSummary'
        - type: object
          properties:
            description:
              type: string
            contact:
              $ref: '#/components/schemas/ContactInfo'
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/Review'
            tips:
              type: array
              items:
                type: string
            nearbyPlaces:
              type: array
              items:
                $ref: '#/components/schemas/PlaceSummary'
    Photo:
      type: object
      properties:
        url:
          type: string
          format: uri
        width:
          type: integer
        height:
          type: integer
    OpeningHours:
      type: object
      properties:
        isOpenNow:
          type: boolean
        periods:
          type: array
          items:
            type: object
            properties:
              open:
                type: object
                properties:
                  day:
                    type: integer
                    minimum: 0
                    maximum: 6
                  time:
                    type: string
              close:
                type: object
                properties:
                  day:
                    type: integer
                    minimum: 0
                    maximum: 6
                  time:
                    type: string
        weekdayText:
          type: array
          items:
            type: string
    ContactInfo:
      type: object
      properties:
        phone:
          type: string
        website:
          type: string
          format: uri
    Review:
      type: object
      properties:
        author:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        text:
          type: string
        time:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            data: null
            error:
              code: "ERR_VALIDATION"
              message: "Validation failed"
              details:
                - field: "startDate"
                  message: "Start date must be in the future"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            data: null
            error:
              code: "ERR_UNAUTHORIZED"
              message: "Authentication required"
...
    NotFound:
      description: Not found - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            data: null
            error:
              code: "ERR_NOT_FOUND"
              message: "Resource not found"
...
    TooManyRequests:
      description: Too many requests - Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            data: null
            error:
              code: "ERR_RATE_LIMIT"
              message: "Rate limit exceeded"
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            data: null
            error:
              code: "ERR_INTERNAL"
              message: "An unexpected error occurred"